# MAINT: Further removal of PyArray_Item_INCREF use