This follows the same pattern as gh-23248 for put and putmask.  (There are a few more, but the pattern should be a bit less common.)

One benchmark gets slower, because it ends up doing no copy at all, but now has to do the setup work anyway, the "dense" is massively faster for the same reason of course:
```
       before           after         ratio
     [b657a75c]       [3a33b4c2]
     <main>           <further-removals>
+        1.40±0μs      1.69±0.01μs     1.21  bench_itemselection.PutMask.time_sparse(False, 'i,O')
+     1.40±0.01μs         1.69±0μs     1.21  bench_itemselection.PutMask.time_sparse(True, 'i,O')
-        5.54±0μs      4.36±0.01μs     0.79  bench_itemselection.PutMask.time_dense(True, 'O')
-     5.63±0.09μs         4.35±0μs     0.77  bench_itemselection.PutMask.time_dense(False, 'O')
-     6.08±0.04μs         4.57±0μs     0.75  bench_itemselection.Put.time_ordered(False, 'O')
-     6.40±0.08μs       4.61±0.2μs     0.72  bench_itemselection.Put.time_ordered(True, 'O')
-         175±4μs      11.3±0.01μs     0.06  bench_itemselection.Put.time_ordered(False, 'i,O')
-         183±3μs       11.7±0.3μs     0.06  bench_itemselection.Put.time_ordered(True, 'i,O')
-       176±0.2μs      10.4±0.02μs     0.06  bench_itemselection.PutMask.time_dense(True, 'i,O')
-         178±3μs      10.4±0.01μs     0.06  bench_itemselection.PutMask.time_dense(False, 'i,O')

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
```