@seberg 2023-02-21
To some degree the refcheck is not quite right for structured dtypes.  I don't want to cross that bridge here, but I think we would need a "trivial copy" flag (problem, while we have nonsense flags, we have no clear flag space left on the dtype instance).